+++ Cargo.toml
@@ -5,6 +5,12 @@ authors = ["mohanson <mohanson@outlook.com>"]
 edition = "2018"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+[lib]
+path = "src/lib.rs"
+
+[[bin]]
+name = "ckb-vm-pprof"
+path = "src/main.rs"
 
 [dependencies]
 addr2line = "0.13.0"
